app/requirements.txt
aiogram==3.5.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic-settings==2.2.1